requirements.txt
aiohttp==2.3.3
amqp==2.2.2
async-timeout==2.0.0
asyncio==3.4.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
enum34==1.1.6
eventlet==0.22.0
gevent==1.2.2
greenlet==0.4.12
idna==2.6
kombu==4.1.0
multidict==3.3.2
python-engineio==2.0.1
python-socketio==1.8.3
pytz==2018.3
redis==2.10.6
requests==2.18.4
six==1.11.0
socketIO-client==0.7.2
urllib3==1.22
vine==1.1.4
websocket-client==0.44.0
yarl==0.14.2
raven==6.6.0